+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/XMLSourceProcessor.java
@@ -17,13 +17,11 @@ package com.liferay.source.formatter;
 import com.liferay.portal.kernel.io.unsync.UnsyncBufferedReader;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringReader;
 import com.liferay.portal.kernel.util.CharPool;
-import com.liferay.portal.kernel.util.NaturalOrderStringComparator;
 import com.liferay.portal.kernel.util.PropertiesUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.tools.ImportPackage;
 import com.liferay.source.formatter.checks.FileCheck;
 import com.liferay.source.formatter.checks.XMLBuildFileCheck;
 import com.liferay.source.formatter.checks.XMLCustomSQLFileCheck;
@@ -40,6 +38,7 @@ import com.liferay.source.formatter.checks.XMLPoshiFileCheck;
 import com.liferay.source.formatter.checks.XMLResourceActionsFileCheck;
 import com.liferay.source.formatter.checks.XMLServiceFileCheck;
 import com.liferay.source.formatter.checks.XMLSolrSchemaFileCheck;
+import com.liferay.source.formatter.checks.XMLSpringFileCheck;
 import com.liferay.source.formatter.checks.XMLWhitespaceCheck;
 import com.liferay.source.formatter.util.FileUtil;
 import com.liferay.util.ContentUtil;
@@ -254,11 +253,8 @@ public class XMLSourceProcessor extends BaseSourceProcessor {
 
 		String newContent = content;
 
-		if (fileName.endsWith("-spring.xml")) {
-			formatSpringXML(fileName, newContent);
-		}
-		else if ((portalSource || subrepository) &&
-				 fileName.endsWith("/struts-config.xml")) {
+		if ((portalSource || subrepository) &&
+			fileName.endsWith("/struts-config.xml")) {
 
 			formatStrutsConfigXML(fileName, newContent);
 		}
@@ -305,16 +301,6 @@ public class XMLSourceProcessor extends BaseSourceProcessor {
 	@Override
 	protected String[] doGetIncludes() {
 		return _INCLUDES;
-	}
-
-	protected void formatSpringXML(String fileName, String content)
-		throws Exception {
-
-		Document document = readXML(content);
-
-		checkOrder(
-			fileName, document.getRootElement(), "bean", null,
-			new SpringBeanElementComparator("id"));
 	}
 
 	protected void formatStrutsConfigXML(String fileName, String content)
@@ -491,6 +477,7 @@ public class XMLSourceProcessor extends BaseSourceProcessor {
 				_serviceFinderColumnSortExcludes, portalSource, subrepository,
 				_portalTablesContent, _pluginsInsideModulesDirectoryNames));
 		fileChecks.add(new XMLSolrSchemaFileCheck());
+		fileChecks.add(new XMLSpringFileCheck());
 
 		fileChecks.add(
 			new XMLBuildFileCheck(sourceFormatterArgs.getBaseDirName()));
@@ -625,163 +612,6 @@ public class XMLSourceProcessor extends BaseSourceProcessor {
 	private List<String> _pluginsInsideModulesDirectoryNames;
 	private String _portalTablesContent;
 	private List<String> _serviceFinderColumnSortExcludes;
-
-	private class SpringBeanElementComparator extends ElementComparator {
-
-		public SpringBeanElementComparator(String nameAttribute) {
-			super(nameAttribute);
-		}
-
-		@Override
-		public int compare(Element element1, Element element2) {
-			String elementName1 = getElementName(element1);
-			String elementName2 = getElementName(element2);
-
-			if ((elementName1 == null) || (elementName2 == null)) {
-				return 0;
-			}
-
-			int startsWithWeight = StringUtil.startsWithWeight(
-				elementName1, elementName2);
-
-			if (startsWithWeight != 0) {
-				String startKey = elementName1.substring(0, startsWithWeight);
-
-				if (startKey.contains(".service.")) {
-					return _compareServiceElements(elementName1, elementName2);
-				}
-			}
-
-			if ((StringUtil.count(elementName1, StringPool.PERIOD) > 1) &&
-				(StringUtil.count(elementName2, StringPool.PERIOD) > 1)) {
-
-				ImportPackage importPackage1 = new ImportPackage(
-					elementName1, false, elementName1);
-				ImportPackage importPackage2 = new ImportPackage(
-					elementName2, false, elementName2);
-
-				return importPackage1.compareTo(importPackage2);
-			}
-
-			if (StringUtil.count(elementName1, StringPool.PERIOD) > 1) {
-				return -1;
-			}
-
-			return super.compare(element1, element2);
-		}
-
-		@Override
-		public String getElementName(Element element) {
-			String elementName = super.getElementName(element);
-
-			if ((elementName != null) &&
-				(StringUtil.count(elementName, StringPool.PERIOD) > 1)) {
-
-				return elementName;
-			}
-
-			return element.attributeValue("class");
-		}
-
-		private int _compareServiceElements(String name1, String name2) {
-			SpringBeanServiceElement springBeanServiceElemen1 =
-				new SpringBeanServiceElement(name1);
-			SpringBeanServiceElement springBeanServiceElemen2 =
-				new SpringBeanServiceElement(name2);
-
-			return springBeanServiceElemen1.compareTo(springBeanServiceElemen2);
-		}
-
-		private class SpringBeanServiceElement
-			implements Comparable<SpringBeanServiceElement> {
-
-			public SpringBeanServiceElement(String name) {
-				_beanObjectName = StringPool.BLANK;
-				_type = -1;
-
-				Matcher matcher = _localServicePattern.matcher(name);
-
-				if (matcher.find()) {
-					_beanObjectName = matcher.group(1);
-					_type = _LOCAL_SERVICE;
-
-					return;
-				}
-
-				matcher = _servicePattern.matcher(name);
-
-				if (matcher.find()) {
-					_beanObjectName = matcher.group(1);
-					_type = _SERVICE;
-
-					return;
-				}
-
-				matcher = _persistencePattern.matcher(name);
-
-				if (matcher.find()) {
-					_beanObjectName = matcher.group(1);
-					_type = _PERSISTENCE;
-
-					return;
-				}
-
-				matcher = _finderPattern.matcher(name);
-
-				if (matcher.find()) {
-					_beanObjectName = matcher.group(1);
-					_type = _FINDER;
-				}
-			}
-
-			@Override
-			public int compareTo(
-				SpringBeanServiceElement springBeanServiceElement) {
-
-				if (_beanObjectName.equals(
-						springBeanServiceElement.getBeanObjectName())) {
-
-					return _type - springBeanServiceElement.getType();
-				}
-
-				NaturalOrderStringComparator comparator =
-					new NaturalOrderStringComparator();
-
-				return comparator.compare(
-					_beanObjectName,
-					springBeanServiceElement.getBeanObjectName());
-			}
-
-			public String getBeanObjectName() {
-				return _beanObjectName;
-			}
-
-			public int getType() {
-				return _type;
-			}
-
-			private static final int _FINDER = 4;
-
-			private static final int _LOCAL_SERVICE = 1;
-
-			private static final int _PERSISTENCE = 3;
-
-			private static final int _SERVICE = 2;
-
-			private String _beanObjectName;
-			private final Pattern _finderPattern = Pattern.compile(
-				"\\.service\\.persistence\\.(\\w+)Finder");
-			private final Pattern _localServicePattern = Pattern.compile(
-				"\\.service\\.(\\w+)LocalService");
-			private final Pattern _persistencePattern = Pattern.compile(
-				"\\.service\\.persistence\\.(\\w+)Persistence");
-			private final Pattern _servicePattern = Pattern.compile(
-				"\\.service\\.(\\w+)Service");
-			private int _type;
-
-		}
-
-	}
 
 	private class StrutsActionElementComparator extends ElementComparator {
